requirements.txt
PyQt6>=6.0.0
numpy>=1.20.0
simpleaudio>=1.0.4
pillow>=10.0.0